+++ src/TheFoundation/main.py
@@ -1,61 +1,40 @@
 import gc
 import json
 
-from libs.webapp import asyncio, HTTP, WebSocket, Schedule, RouterHTTP
+from libs.thefoundation import asyncio, TheFoundation, Request, WebSocket, Schedule
 
 
-app = RouterHTTP()
+app = TheFoundation()
 
 for crendential in json.load(open('/credentials.json')):
-    if app.setup(crendential['ssid'], crendential['password'], timezone=2):
+    if app.connect(crendential['ssid'], crendential['password'], timezone=2):
         break
-
-if not app.is_connected:
+else:
     raise RuntimeError('Connection failed to WiFi')
+
 
 app.mount(path='/www/public', name='/public')
 
-@app.http(methods='GET|POST', pattern='/')
-async def c(http: HTTP) -> int:
-    status_code, content = http.response.template('www/templates/ws.html')
-    http.response.content = content
+@app.map(methods='GET', pattern='/')
+async def a(request: Request):
+    return request.redirect(to='/ws')
 
-    # async def backg_task():
-    #     for i in range(100):
-    #         print('bg:', i)
-    #         await asyncio.sleep(0.2)
+@app.map(methods='GET', pattern='/download')
+async def b(request: Request):
+    return request.attachment('/www/public/favicon.png')
 
-    # http.add_background_task(backg_task)
+@app.map(methods='GET', pattern='/ws')
+async def c(request: Request):
+    return request.template('www/templates/ws.html')
 
-    return status_code
-
-@app.http(methods='GET|POST', pattern='/templating')
-async def d(http: HTTP) -> int:
-    status_code, content = http.response.template('www/templates/index.html', {
-        'metadata': {
-            'uuid_0': {
-                'title': 'Hello World 1!',
-                'description': 'Hello my world 1!'
-            },
-            'uuid_1': {
-                'title': 'Hello World 2!',
-                'description': 'Hello my world 2!'
-            },
-            'uuid_2': {
-                'title': 'Hello World 3!',
-                'description': 'Hello my world 3!'
-            }
-        },
+@app.map(methods='GET', pattern='/json')
+async def d(request: Request):
+    return request.json({
+        'app': 'TheFoundation',
+        'version': 0.2
     })
 
-    http.response.content = content
-    return status_code
-
-@app.http(methods='GET|POST', pattern='/download')
-async def d(http: HTTP) -> int:
-    return http.response.attachment('www/public/faviscon.png')
-
-@app.websocket(port=8000, max_tasks=2)
+@app.websocket(port=8000, max_connections=2)
 async def e(websocket: WebSocket):
     while True:
         try:
@@ -67,16 +46,26 @@ async def e(websocket: WebSocket):
             break
         except Exception as e:
             print(str(e))
+            break
 
-@app.schedule(hour=3)
+@app.websocket(port=8001, max_connections=2)
+async def e(websocket: WebSocket):
+    while True:
+        try:
+            message = await websocket.recv()
+            # print('From:', websocket.address, ' - ', message)
+            # websocket.send(str(websocket))
+            websocket.broadcast(str(websocket.websockets))
+        except WebSocket.WebSocketDisconnect:
+            break
+        except Exception as e:
+            print(str(e))
+            break
+
+@app.schedule(second=[0, 15, 30, 45])
 async def f(schedule: Schedule):
     print('Triggered:', schedule.name, 'at:', schedule.localtime)
-    schedule.stop()
-    await asyncio.sleep(0)
-
-@app.schedule(second=0)
-async def f(schedule: Schedule):
-    print('Free memory:', round(gc.mem_free() / 1024, 2), 'kb.')
+    print('Free memory:', round(gc.mem_free() / 1024, 2), 'kb.', schedule.n)
     await asyncio.sleep(0)
 
 app.listen()
